+++ cli/cli/commands/service/service.go
@@ -26,7 +26,7 @@ func init() {
 	ServiceCmd.AddCommand(add.ServiceAddCmd.MustGetCobraCommand())
 	ServiceCmd.AddCommand(logs.ServiceLogsCmd.MustGetCobraCommand())
 	ServiceCmd.AddCommand(rm.ServiceRmCmd.MustGetCobraCommand())
-	ServiceCmd.AddCommand(shell.ShellCmd)
+	ServiceCmd.AddCommand(shell.ServiceShellCmd.MustGetCobraCommand())
 	ServiceCmd.AddCommand(pause.PauseCmd.MustGetCobraCommand())
 	ServiceCmd.AddCommand(unpause.UnpauseCmd.MustGetCobraCommand())
 }